reorganizes index page commands
