+++ cypress/support/index.js
@@ -14,6 +14,7 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
+import './index-page.commands.js';
 import './commands'
 
 // Alternatively you can use CommonJS syntax:
